Use correct variable when setting content length (#72)
None is stringable so the previous code here would work, but, in
theory could eventually lead to "badness". The intent of the code
was to match on a USize | None and set the content length to the
string version of the USize if a USize.

However, it was setting to the string version of USize | None.
Given the match statement, this was fine for now, but, the intent
was clouded.

This could lead to a bug in the future if someone changed the code.